requirements.txt
flask
pymysql
flask-bcrypt
cryptography
python-dotenv
mysql-connector-python